continue repo foundational setup, reorganize dashboard code, add types to synthetix js, and start working on styling dashboard
