+++ sections/dashboard/index.ts
@@ -1,0 +1,5 @@
+import StatBoxes from './StatBoxes';
+import BarStats from './BarStats';
+import PossibleActions from './PossibleActions';
+
+export { StatBoxes, BarStats, PossibleActions };
